Updating todo id to be DateTime.Now.Ticks to ensure uniqueness

+++ Ianf.Global360ToDo.Domain/ToDo.cs
@@ -3,7 +3,7 @@
 public readonly record struct ToDo
 {
     public ToDo() { }
-    public int Id { get; init; } = 0;
+    public long Id { get; init; } = 0;
     public string Title { get; init; } = string.Empty;
     public string Contents { get; init; } = string.Empty;
 }
