+++ Ianf.Global360ToDo.Services/IToDoService.cs
@@ -4,7 +4,7 @@ namespace Ianf.Global360ToDo.Services;
 
 public interface IToDoService
 {
-    public Task<int> AddToDoItem(NewToDo newToDo);
-    public Task<List<ToDo>> RemoveToDoItem(int toDoId);
+    public Task<long> AddToDoItem(NewToDo newToDo);
+    public Task<List<ToDo>> RemoveToDoItem(long toDoId);
     public Task<List<ToDo>> GetToDoItems();
 }
